requirements.txt
# https://pypi.org/project/Flask
Flask==3.0.2
pandas
gspread

gspread-dataframe

google-api-python-client
